# frontend/src/api/generated/openapi.json
{
  "components": {
    "schemas": {
...
      "CandidateGroupsCommand": {
        "description": "归类联动动作的请求体：本次动作涉及的观察组。\n\n缺失与显式 null 都表示没选组（同迁移前），交给服务判定成 400，不在这里收紧成必填；\n非数组取值与含非字符串的数组由结构校验拒绝，取代迁移前「把字符串逐字符拆成组标识」\n与 `str()` 强转的宽松转换。",
        "properties": {
          "groupIds": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Groupids"
          }
        },
        "title": "CandidateGroupsCommand",
        "type": "object"
      },
...
      "DeleteGroupResponse": {
        "description": "删除组成功只回执被删的组。",
        "properties": {
          "deleted": {
            "title": "Deleted",
            "type": "string"
          }
        },
        "required": [
          "deleted"
        ],
        "title": "DeleteGroupResponse",
        "type": "object"
      },
...
      "FailurePayload": {
        "description": "业务失败响应：哪个状态码适用仍由各端点自己声明（`responses`）。",
        "properties": {
          "detail": {
            "$ref": "#/components/schemas/FailureDetail"
          }
        },
        "required": [
          "detail"
        ],
        "title": "FailurePayload",
        "type": "object"
      },
      "GroupListPayload": {
        "description": "组列表：组名与成员数量来自服务端。",
        "properties": {
          "groups": {
            "items": {
              "$ref": "#/components/schemas/ObservationGroupPayload"
            },
            "title": "Groups",
            "type": "array"
          }
        },
        "required": [
          "groups"
        ],
        "title": "GroupListPayload",
        "type": "object"
      },
      "GroupNameCommand": {
        "description": "新建与重命名观察组的请求体：只认识组名。\n\n空名、纯空白与超长仍由领域校验拒绝成 400；组名是字符串线格式，\n因此非字符串取值由结构校验拒绝，不再先 `str()` 成字符串再建组。",
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          }
        },
        "title": "GroupNameCommand",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ObservationGroupPayload": {
        "description": "观察组：只有创建、重命名与删除，没有归档。",
        "properties": {
          "groupId": {
            "title": "Groupid",
            "type": "string"
          },
          "isDefault": {
            "title": "Isdefault",
            "type": "boolean"
          },
          "memberCount": {
            "title": "Membercount",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "groupId",
          "name",
          "isDefault",
          "memberCount"
        ],
        "title": "ObservationGroupPayload",
        "type": "object"
      },
      "ObservationSavePayload": {
        "description": "归类联动动作结果：有效关系、写入后的候选项与同一份写入版本。\n\n与候选归类的动作响应同一形状：前端据此更新列表并带着写入后的导航修订号发出\n自动推进，因此一次归类联动不需要中间再读一遍完整浏览结果。版本号在顶层与\n候选项上各出现一次是既有线格式（前端用候选项里的那份）；是否收拢留给工单 06。",
        "properties": {
          "candidate": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ObservationCandidatePayload"
              },
              {
                "type": "null"
              }
            ]
          },
          "groupIds": {
            "items": {
              "type": "string"
            },
            "title": "Groupids",
            "type": "array"
          },
          "listRemoved": {
            "items": {
              "type": "string"
            },
            "title": "Listremoved",
            "type": "array"
          },
          "listRevision": {
            "title": "Listrevision",
            "type": "integer"
          },
          "listRevisionBefore": {
            "title": "Listrevisionbefore",
            "type": "integer"
          },
          "navigationRevision": {
            "title": "Navigationrevision",
            "type": "integer"
          }
        },
        "required": [
          "listRevision",
          "listRevisionBefore",
          "listRemoved",
          "navigationRevision",
          "groupIds",
          "candidate"
        ],
        "title": "ObservationSavePayload",
        "type": "object"
      },
...
      "ObservationStockListPayload": {
        "description": "按可选组筛选的观察列表。",
        "properties": {
          "stocks": {
            "items": {
              "$ref": "#/components/schemas/ObservedStockPayload"
            },
            "title": "Stocks",
            "type": "array"
          }
        },
        "required": [
          "stocks"
        ],
        "title": "ObservationStockListPayload",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "DailyScreen Lite",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/observations/candidates/{candidate_id}/observe": {
      "post": {
        "description": "加入或保留观察：增加所选关系并完成本次归类（同一事务）。",
        "operationId": "observe_candidate_api_observations_candidates__candidate_id__observe_post",
        "parameters": [
          {
            "in": "path",
            "name": "candidate_id",
            "required": true,
            "schema": {
              "title": "Candidate Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CandidateGroupsCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObservationSavePayload"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FailurePayload"
                }
              }
            },
            "description": "关系无效或当前状态不允许"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FailurePayload"
                }
              }
            },
            "description": "候选项不存在"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Observe Candidate",
        "tags": [
          "observations"
        ]
      }
    },
    "/api/observations/candidates/{candidate_id}/remove-from-groups": {
      "post": {
        "description": "移出观察组：移除所选关系并同时记为暂不关注（同一事务）。",
        "operationId": "remove_from_groups_api_observations_candidates__candidate_id__remove_from_groups_post",
        "parameters": [
          {
            "in": "path",
            "name": "candidate_id",
            "required": true,
            "schema": {
              "title": "Candidate Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CandidateGroupsCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObservationSavePayload"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FailurePayload"
                }
              }
            },
            "description": "关系无效或当前状态不允许"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FailurePayload"
                }
              }
            },
            "description": "候选项不存在"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Remove From Groups",
        "tags": [
          "observations"
        ]
      }
    },
    "/api/observations/groups": {
      "get": {
        "operationId": "list_groups_api_observations_groups_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupListPayload"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Groups",
        "tags": [
          "observations"
        ]
      },
      "post": {
        "operationId": "create_group_api_observations_groups_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GroupNameCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObservationGroupPayload"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FailurePayload"
                }
              }
            },
            "description": "组名不合规或重名"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Group",
        "tags": [
          "observations"
        ]
      }
    },
    "/api/observations/groups/{group_id}": {
      "delete": {
        "description": "删除组及其成员关系：不删除来源、笔记与处理历史，也不影响其他组。",
        "operationId": "delete_group_api_observations_groups__group_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "group_id",
            "required": true,
            "schema": {
              "title": "Group Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteGroupResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FailurePayload"
                }
              }
            },
            "description": "观察组不存在"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Group",
        "tags": [
          "observations"
        ]
      },
      "patch": {
        "operationId": "rename_group_api_observations_groups__group_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "group_id",
            "required": true,
            "schema": {
              "title": "Group Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GroupNameCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObservationGroupPayload"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FailurePayload"
                }
              }
            },
            "description": "组名不合规或重名"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FailurePayload"
                }
              }
            },
            "description": "观察组不存在"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Rename Group",
        "tags": [
          "observations"
        ]
      }
    },
    "/api/observations/memberships": {
      "get": {
        "description": "某股票当前所属的观察组，用于卡片与详情预选。",
        "operationId": "memberships_api_observations_memberships_get",
        "parameters": [
          {
            "in": "query",
            "name": "securityId",
            "required": true,
            "schema": {
              "title": "Securityid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MembershipPayload"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Memberships",
        "tags": [
          "observations"
        ]
      },
      "put": {
        "description": "整组替换（观察组页面的转组、增减归属、退出全部组），不改变候选处理状态。",
        "operationId": "save_membership_api_observations_memberships_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MembershipCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MembershipPayload"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FailurePayload"
                }
              }
            },
            "description": "缺少证券标识、未完成归类或组不存在"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FailurePayload"
                }
              }
            },
            "description": "股票未经导入与识别"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Save Membership",
        "tags": [
          "observations"
        ]
      }
    },
    "/api/observations/securities": {
      "get": {
        "description": "观察列表：按证券去重，带所属全部现存组与加入时间；未知组明确报 404。",
        "operationId": "list_stocks_api_observations_securities_get",
        "parameters": [
          {
            "in": "query",
            "name": "groupId",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Groupid"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObservationStockListPayload"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FailurePayload"
                }
              }
            },
            "description": "证券或观察关系不存在"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Stocks",
        "tags": [
          "observations"
        ]
      }
    },
    "/api/observations/securities/{security_id}": {
      "get": {
        "description": "共用个股详情：股票信息、观察关系与该证券的行情归属、笔记与处理记录入口。",
        "operationId": "stock_detail_api_observations_securities__security_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "security_id",
            "required": true,
            "schema": {
              "title": "Security Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StockDetailPayload"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FailurePayload"
                }
              }
            },
            "description": "证券或观察关系不存在"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stock Detail",
        "tags": [
          "observations"
        ]
      }
    },
    "/api/observations/securities/{security_id}/focus": {
      "post": {
        "description": "跨模块打开详情：切到汇总视图并把这支股票设为观察模块的当前股票。",
        "operationId": "focus_stock_api_observations_securities__security_id__focus_post",
        "parameters": [
          {
            "in": "path",
            "name": "security_id",
            "required": true,
            "schema": {
              "title": "Security Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObservationViewPayload"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FailurePayload"
                }
              }
            },
            "description": "证券或观察关系不存在"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Focus Stock",
        "tags": [
          "observations"
        ]
      }
    },
    "/api/observations/view": {
      "get": {
        "description": "观察组模块的组、列表与浏览上下文，用于刷新与重启后恢复。",
        "operationId": "get_view_api_observations_view_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObservationViewPayload"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get View",
        "tags": [
          "observations"
        ]
      },
      "put": {
        "description": "更新当前组、排序或当前股票；只影响观察组模块自身的工作位置。",
        "operationId": "update_view_api_observations_view_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ObservationViewChanges"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObservationViewPayload"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FailurePayload"
                }
              }
            },
            "description": "未知排序方式"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update View",
        "tags": [
          "observations"
        ]
      }
    },
    "/api/quotes/summary": {
      "get": {
        "description": "批量轻量行情摘要：队列列表的最新收盘价、行情日与日涨跌幅。\n\n只读取最近两个交易日，不为列表加载整段历史；缺失的股票不返回条目。",
        "operationId": "quote_summaries_api_quotes_summary_get",
        "parameters": [
          {
            "in": "query",
            "name": "securityIds",
            "required": true,
            "schema": {
              "title": "Securityids",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Quote Summaries Api Quotes Summary Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Quote Summaries",
        "tags": [
          "quotes"
        ]
      }
    },
...
  }
}